possible de compiler, mtn voir ce que ca donne et corriger les erreurs

+++ engine/src/ecs/factories/TriggerFactories.cpp
@@ -9,7 +9,7 @@ entt::entity	TriggerFactories::create(
 	float			y,
 	int				w,
 	int				h,
-	int				duration,
+	float			duration,
 	bool			once
 )
 {
